1365 Add new stage to delete tables (#211)
* refactored delete tables stage

* added delete tables stage

* updated delete tables to include docstring

* updated delete tables to include docstring

+++ cishouseholds/pipeline/load.py
@@ -27,11 +27,22 @@ def extract_from_table(table_name: str):
 
 
 @register_pipeline_stage("delete_tables")
-def delete_tables(table_names: str):
+def delete_tables(**kwargs):
+    """
+    All key word args are optional:
+    Specify table_names as a list contaning absolute table names to remove.
+    Specify pattern as a string in SQL format to set a specific SQL pattern to remove.
+    Specify prefix to remove all tables with a given prefix.
+    """
     spark_session = get_or_create_spark_session()
     prefix = get_config()["storage"]["table_prefix"]
-    for table_name in table_names:
-        spark_session.sql(f"DROP TABLE IF EXISTS {prefix}_{table_name}")
+    if "table_names" in kwargs and kwargs["table_names"] != "":
+        for table_name in kwargs["table_names"]:
+            spark_session.sql(f"DROP TABLE IF EXISTS {prefix}_{table_name}")
+    if "pattern" in kwargs and kwargs["pattern"] != "":
+        spark_session.sql(f"DROP TABLE IF EXISTS LIKE {kwargs['pattern']}")
+    if "prefix" in kwargs and kwargs["prefix"] != "":
+        spark_session.sql(f"DROP TABLE IF EXISTS LIKE {kwargs['prefix'].replace('_','~_')}% ESCAPE '~'")
 
 
 def add_error_file_log_entry(file_path: str, error_text: str):
